packages/ai-models/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prometheus-ai"
version = "0.1.0"
description = "Prometheus AI - Mining Pool Prediction and Optimization Engine"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Viddhana Pool Team", email = "ai@viddhana.io" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "mining", "prediction", "optimization", "lstm", "anomaly-detection"]

dependencies = [
    # Deep Learning
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    
    # Time Series & Anomaly Detection
    "neuralprophet>=0.7.0",
    "pyod>=1.1.0",
    
    # API Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Data Storage
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    
    # MLOps
    "mlflow>=2.9.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "structlog>=23.2.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.11.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
prometheus-serve = "prometheus.serving.api:main"
prometheus-train = "prometheus.training.trainer:main"

[project.urls]
Homepage = "https://github.com/viddhana-pool/prometheus-ai"
Documentation = "https://docs.viddhana.io/prometheus"
Repository = "https://github.com/viddhana-pool/prometheus-ai"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.isort]
known-first-party = ["prometheus"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=prometheus --cov-report=term-missing"
